Add files via upload
Implementation of a feed-forward neural network (NN) that performs classification on the IRIS dataset. The Neural Network will be trained via the stochastic gradient method (SGD), forward propagation and back propagation algorithms. The performance of the implementation is compared with that of the built-in Neural Network module available in scikit-learn